COMP: fix warning

+++ Source/cmGlobalGenerator.cxx
@@ -503,7 +503,7 @@ void cmGlobalGenerator::Configure()
   // update the cache entry for the number of local generators, this is used
   // for progress
   char num[100];
-  sprintf(num,"%d",m_LocalGenerators.size());
+  sprintf(num,"%d",static_cast<int>(m_LocalGenerators.size()));
   this->GetCMakeInstance()->AddCacheEntry
     ("CMAKE_NUMBER_OF_LOCAL_GENERATORS", num,
      "number of local generators",
